add lazy measures
It isn't always the case that a language needs to have the entire
measure computed. For example, in Disrupt there is no need to compute
the probability of the sentinel value corresponding to the observation
being false.

Now, `infer` includes an option for lazy measures to compute the measure
only as needed. Using this option causes `support` to be approximate
because some points may actually have zero measure due to laziness.
However, it should always be a superset of the actual support.